+++ src/app.module.ts
@@ -24,6 +24,7 @@ import { ReadingProgressModule } from './modules/reading-progress/reading-progre
 import { ViewStatsModule } from './modules/view-stats/view-stats.module';
 import { UploadsModule } from './modules/uploads/uploads.module';
 import { StatusModule } from './modules/status/status.module';
+import { TagsModule } from './modules/tags/tags.module';
 
 const staticRoot = join(process.cwd(), process.env.LOCAL_UPLOADS_DIR ?? 'var/uploads');
 
@@ -53,6 +54,7 @@ const staticRoot = join(process.cwd(), process.env.LOCAL_UPLOADS_DIR ?? 'var/upl
     ViewStatsModule,
     UploadsModule,
     StatusModule,
+    TagsModule,
     // ...другие модули
   ],
   controllers: [AppController],