+++ src/modules/tags/tags.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { PrismaService } from '../../prisma/prisma.service';
+import { TagsService } from './tags.service';
+import { TagsController } from './tags.controller';
+
+@Module({
+  controllers: [TagsController],
+  providers: [PrismaService, TagsService],
+  exports: [TagsService],
+})
+export class TagsModule {}